+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
         * [HQX-269] - Initial work on bulk removal of clients from groups
         * [HQX-370] - Add functionality to display and update client credit score tenure on client profile
         * [HQX-358] - Approval flow: Removal of farmers from groups
+        * [HQX-332/HQX-379] - Add markedForCreditScoreRefresh and allowedEnrolmentsPerClient fields on loan product UI
 
 ## Version 1.4.11 - Community 1.0.0