+++ RELEASE_NOTES.md
@@ -9,6 +9,7 @@
         * [HQX-370] - Add functionality to display and update client credit score tenure on client profile
         * [HQX-358] - Approval flow: Removal of farmers from groups
         * [HQX-332/HQX-379] - Add markedForCreditScoreRefresh and allowedEnrolmentsPerClient fields on loan product UI
+        * [HQX-402] - Guard bulk removal of clients from groups using permissions and filter groups by offices selected
 
 ## Version 1.4.11 - Community 1.0.0